refactor(all): remove complexity and unused expressions from nix-config
Currently only supporting `flex5i` host, however, plan to support more since now the architecture is overall simpler.

Also, I'm using gnome for now (was having some problems using hyprland compositor based on wayland) but I intend to rice a xmonad using the current archictecture.

Main difference is the explicitly, now the configuration does not rely on one big abstraction (mkMachine function before), and rely on smaller, but simpler, powerful and concise abstractions, allowing the granular adjustment.

Another big difference is the better use of nix flakes.

To conclude, it's been decided to remove all unnecessary Nix modules created before, and the better use/reliance of nixpkgs existing modules.

To create a new module the following questions need to be asked:
- Is it already easy to do that thing using existing modules in nixpkgs?
- Does it interact with many aspects of the system? (e.g., user groups, home-manager modules, interactivity w/ other services).
- Does it hide too much important configuration that needs to be explicit?

Them created system modules can use home-manager, but needs to exist an explicit option asking if it will use the home-manager (if its for ricing) like the shell.zsh module, if not for ricing, the module also can use home-manager if depends on that (like the direnv module), but overall this pattern is discouraged since home-manager modules should belong to it's own module system.